+++ e2e/compare/output/console.js
@@ -4,6 +4,10 @@ const printRegularLine = (entry) => {
     console.debug(` - ${entry.name}: ${formatDurationDiffChange(entry)}`);
 };
 
+/**
+ * Prints the result simply to console.
+ * @param {Object} data
+ */
 module.exports = (data) => {
     // No need to log errors or warnings as these were be logged on the fly
     console.debug('');
