+++ e2e/server/commands.js
@@ -1,3 +1,7 @@
+/**
+ * Type name for commands that are understood among the app/client and the server.
+ * @type {{LOGOUT: string, WAIT_FOR_APP_READY: string, LOGIN: string, REQUEST_PERFORMANCE_METRICS: string}}
+ */
 module.exports = {
     LOGIN: 'login',
     LOGOUT: 'logout',
