+++ e2e/server/utils/tryParseJSON.js
@@ -1,0 +1,12 @@
+const tryParseJSON = (jsonString) => {
+    try {
+        const o = JSON.parse(jsonString);
+        if (o && typeof o === 'object') {
+            return o;
+        }
+    } catch (e) {
+        // ignore
+    }
+};
+
+module.exports = tryParseJSON;
