+++ e2e/server/utils/withFailTimeout.js
@@ -1,0 +1,10 @@
+const {INTERACTION_TIMEOUT} = require('../../config');
+
+const withFailTimeout = (promise, name) => {
+    const timeoutId = setTimeout(() => {
+        throw new Error(`[${name}] Interaction timed out`);
+    }, INTERACTION_TIMEOUT);
+    return promise.finally(() => clearTimeout(timeoutId));
+};
+
+module.exports = withFailTimeout;
